refactor: refactor cep request for mvc structure
