+++ src/model/service/ViaCEPClient.java
@@ -1,4 +1,4 @@
-package service;
+package model.service;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,7 +6,7 @@ import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 
-import service.exception.ViaCEPException;
+import model.exceptions.ViaCEPException;
 
 public class ViaCEPClient {
 	private static final String BASE_URL = "https://viacep.com.br/ws/";
@@ -28,7 +28,7 @@ public class ViaCEPClient {
 				}
 
 				if (response.toString().contains("\"erro\": true")) {
-					throw new ViaCEPException("CEP not recognized.");
+					throw new ViaCEPException("CEP inválido. Insira um CEP válido para continuar.");
 				}
 
 				reader.close();
@@ -36,10 +36,10 @@ public class ViaCEPClient {
 
 				return response.toString();
 			} else {
-				throw new ViaCEPException("Failed to connect. Error code: " + connection.getResponseCode());
+				throw new ViaCEPException("Falha durante a conexão. Conecte-se a internet para continuar o cadastro.");
 			}
 		} catch (IOException e) {
-			throw new ViaCEPException("Error while accessing ViaCEP API: " + e.getMessage(), e);
+			throw new IOException("Informe um CEP válido para continuar.");
 		}
 	}
 }
